refactor(markdown-composer): restructure into tab-ready layout
Move the existing Composer screen into a composer/ subfolder behind a
new thin tab-shell parent, matching the base64/json tab pattern, so a
second tab (Markdown Editor) can be added without reworking routing.

Closes #146
