+++ src/routes/router.tsx
@@ -28,7 +28,8 @@ import { JsonFormatter } from "~/screens/json/formatter/json-formatter";
 import { Json } from "~/screens/json/json";
 import { JsonRepair } from "~/screens/json/repair/json-repair";
 import { JwtDecoder } from "~/screens/jwt-decoder/jwt-decoder";
-import { MarkdownComposerScreen } from "~/screens/markdown-composer/markdown-composer";
+import { MarkdownComposerScreen } from "~/screens/markdown-composer/composer/composer";
+import { MarkdownComposer } from "~/screens/markdown-composer/markdown-composer";
 import { PokerPlanning } from "~/screens/poker-planning/poker-planning";
 import { QrcodeDecoder } from "~/screens/qrcode/decoder/qrcode-decoder";
 import { QrcodeGenerator } from "~/screens/qrcode/generator/qrcode-generator";
@@ -324,10 +325,31 @@ const diffViewerRoute = createRoute({
 const markdownComposerRoute = createRoute({
   getParentRoute: () => rootRoute,
   path: "/markdown-composer",
+  component: MarkdownComposer,
+});
+
+const markdownComposerIndexRoute = createRoute({
+  getParentRoute: () => markdownComposerRoute,
+  path: "/",
+  component: () => <Navigate to="/markdown-composer/composer" replace />,
+});
+
+const markdownComposerComposerRoute = createRoute({
+  getParentRoute: () => markdownComposerRoute,
+  path: "/composer",
   component: MarkdownComposerScreen,
 });
 
-export const ROUTES_WITH_CHILDREN = ["/url", "/base64", "/json", "/colors", "/common-lists", "/qrcode", "/image-ocr"];
+export const ROUTES_WITH_CHILDREN = [
+  "/url",
+  "/base64",
+  "/json",
+  "/colors",
+  "/common-lists",
+  "/qrcode",
+  "/image-ocr",
+  "/markdown-composer",
+];
 
 const routeTree = rootRoute.addChildren([
   homeRoute,
@@ -356,7 +378,7 @@ const routeTree = rootRoute.addChildren([
   vr3dViewerRoute,
   codingStandardsRoute,
   diffViewerRoute,
-  markdownComposerRoute,
+  markdownComposerRoute.addChildren([markdownComposerIndexRoute, markdownComposerComposerRoute]),
 ]);
 
 const hashHistory = createHashHistory();
